requirements.txt
Flask==3.1.2
python-telegram-bot==22.6
python-dotenv==1.2.1